+++ src/com/ui/HelloWorldClass.java
@@ -1,4 +1,4 @@
-package ui;
+package com.ui;
 
 import javax.swing.*;
 import java.awt.event.*;
