+++ src/com/ui/JavaUI.java
@@ -1,4 +1,4 @@
-package ui;
+package com.ui;
 
 import javax.swing.*;
